fix(Phoenix.Installer): ensure es2015 preset works out of the box
There are two issues with brunch as it stands. Out of the box there is
an error:

 > error: Compiling of 'web/static/js/app.js' failed. Couldn't find
   preset "es2015"

To fix this error the `babel-preset-es2015` package must be available.

After the package is available, a new error appears due to babel
attempting to compile a package in the node_modules directory.

This is fixed by removing `npm: true` from the `brunch-config.json`
file.